fixed error with fitter
The newly updated version of the fitter package gets rid of the verbose argument so I updated the code accordingly. I ran the final notebooks and they ran with no problems. I am yet to run @hassan-obeid notebooks because I am unclear on which ones will be final.

@timothyb0912  any thoughts on how to prevent this type of issue in the future since we rely on third party libraries?
I also changed the version number for black in the environment file since pip wasn't able to find a version 19.10